requirements.txt
APScheduler==3.6.3
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.10.0
cachetools==3.1.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
enum34==1.1.10
et-xmlfile==1.1.0
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
h11==0.12.0
httplib2==0.13.1
idna==3.3
oauth2client==4.1.3
oauthlib==3.1.1
outcome==1.1.0
protobuf==3.19.0
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycparser==2.20
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyTelegramBotAPI==4.1.1
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.0
selenium==4.0.0
six==1.12.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
tornado==6.1
trio==0.19.0
trio-websocket==0.9.2
tzdata==2021.2.post0
tzlocal==3.0
uritemplate==3.0.0
urllib3==1.26.7
vk-api==11.9.4
wsproto==1.0.0
